fixed value conversions, introduced virtual copy pattern to fitness class

+++ zad3/include/utils.hpp
@@ -23,8 +23,8 @@ int getRandomInt();
 double getRandomDouble();
 double getRandomGaussian();
 bool willMutate(double mutationRate);
-long averageDistance(Point *point, std::vector<Point *> otherPoints);
-long averageDistance(Cluster *cluster);
-long distanceOfCenters(Cluster *first, Cluster *second);
+double averageDistance(Point *point, std::vector<Point *> otherPoints);
+double averageDistance(Cluster *cluster);
+double distanceOfCenters(Cluster *first, Cluster *second);
 
 std::vector<Point *> readCsv(std::string filepath);
